recipe8-4 end. start no yet done

+++ chapter8/recipe8-4/end/football/src/main/java/com/packt/football/domain/Card.java
@@ -2,6 +2,6 @@ package com.packt.football.domain;
 
 import java.util.Optional;
 
-public record Card(Integer id, Optional<Album> album, Player player) {
+public record Card(Long id, Optional<Album> album, Player player) {
 
 }
